requirements.txt
# Core LlamaIndex modules (modular architecture)
llama-index-core>=0.10.55
llama-index-vector-stores-qdrant>=0.1.5
llama-index-embeddings-huggingface>=0.1.5
llama-index-llms-openai>=0.10.55
# Vector database
qdrant-client>=1.11.0

# Document processing
sentence-transformers>=3.0.0
EbookLib>=0.18
html2text>=2020.1.16
pypdf>=4.2.0
requests>=2.31.0
python-dotenv>=1.0.1
pdfplumber
tiktoken

# embedding model
sentence-transformers


# env
python-dotenv>=1.0.1